+++ server/lib/kernel/src/api/context.rs
@@ -313,13 +313,13 @@ impl BufferManager for StubBufferManager {
     fn get(
         &self,
         _id: crate::mm::BufferId,
-    ) -> Option<Arc<RwLock<dyn crate::api::storage_ops::KernelBuffer>>> {
+    ) -> Option<Arc<RwLock<dyn crate::block::KernelBuffer>>> {
         None
     }
 
     fn register(
         &self,
-        buffer: Arc<RwLock<dyn crate::api::storage_ops::KernelBuffer>>,
+        buffer: Arc<RwLock<dyn crate::block::KernelBuffer>>,
     ) -> crate::mm::BufferId {
         buffer.read().id()
     }
@@ -327,7 +327,7 @@ impl BufferManager for StubBufferManager {
     fn unregister(
         &self,
         _id: crate::mm::BufferId,
-    ) -> Option<Arc<RwLock<dyn crate::api::storage_ops::KernelBuffer>>> {
+    ) -> Option<Arc<RwLock<dyn crate::block::KernelBuffer>>> {
         None
     }