+++ server/modules/codec-utf8/src/domain_tests.rs
@@ -2,7 +2,7 @@
 
 use {
     reovim_driver_codec::{CodecMetadata, ContentType, Decode, Encode, Index},
-    reovim_driver_vfs::ByteEdit,
+    reovim_kernel::api::v1::ByteEdit,
     reovim_types_text::{Text, TextEdit, TextPosition},
 };